+++ packages/google_maps_flutter/example/lib/main.dart
@@ -4,6 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'animate_camera.dart';
+import 'map_coordinates.dart';
 import 'map_ui.dart';
 import 'move_camera.dart';
 import 'page.dart';
@@ -13,6 +14,7 @@ import 'scrolling_map.dart';
 
 final List<Page> _allPages = <Page>[
   MapUiPage(),
+  MapCoordinatesPage(),
   AnimateCameraPage(),
   MoveCameraPage(),
   PlaceMarkerPage(),